+++ CHANGELOG.rst
@@ -2,6 +2,13 @@ Changelog
 #########
 
 
+3.0.0
+*****
+
+- fixed a crash in ``PhpHighlighter::file()`` and ``PhpHighlighter::code()``
+  - these methods now return ``null`` in case of failure
+
+
 2.0.0
 *****